fix all comments
1. Move tests to new file, fix NUnit pacages
2. Create own parser for every tag
