+++ cs/Markdown/Markdown.csproj
@@ -5,9 +5,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="FluentAssertions" Version="5.10.3" />
       <PackageReference Include="NUnit" Version="3.12.0" />
-      <PackageReference Include="NUnit3TestAdapter" Version="3.17.0" />
     </ItemGroup>
 
 </Project>
