+++ cs/MarkdownTest/MarkdownParserTests/ParseHeaderTagTest.cs
@@ -1,4 +1,5 @@
 ﻿using FluentAssertions;
+using Markdown.Parsers;
 using Markdown.Tags;
 using Markdown.Tags.HeaderTag;
 using NUnit.Framework;
@@ -10,13 +11,13 @@ namespace Markdown.MarkdownParserTests
         [Test]
         public void ReadHeaderToken_ShouldReturnEmptyArray_OnSimpleString()
         {
-            MarkdownParser.ParseHeaderTag("asdf").Should().BeEmpty();
+            HeaderParser.ParseTags("asdf").Should().BeEmpty();
         }
 
         [Test]
         public void ReadHeaderToken_ShouldReturnCorrectArray_OnHeaderSting()
         {
-            var result = MarkdownParser.ParseHeaderTag("#asdf");
+            var result = HeaderParser.ParseTags("#asdf");
             result.Length.Should().Be(2);
             result[0].Should().BeEquivalentTo(new OpenHeaderTag(0));
             result[1].Should().BeEquivalentTo(new CloseHeaderTag(5));
