+++ cs/clean-code.sln
@@ -11,6 +11,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Samples", "Samples\Samples.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Markdown", "Markdown\Markdown.csproj", "{C115A15A-9A74-4A25-8A0F-6C2F76F2A4C7}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MarkdownTest", "MarkdownTest\MarkdownTest.csproj", "{4BE2D276-30D9-4053-A5AD-DB54A14E7E8B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,5 +35,9 @@ Global
 		{C115A15A-9A74-4A25-8A0F-6C2F76F2A4C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C115A15A-9A74-4A25-8A0F-6C2F76F2A4C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C115A15A-9A74-4A25-8A0F-6C2F76F2A4C7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4BE2D276-30D9-4053-A5AD-DB54A14E7E8B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4BE2D276-30D9-4053-A5AD-DB54A14E7E8B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4BE2D276-30D9-4053-A5AD-DB54A14E7E8B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4BE2D276-30D9-4053-A5AD-DB54A14E7E8B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
